[feat #58] impement virtual view
set maxsize of document to 750k pixels since even on 6Million pixel in
height we got some rubberband. Also since we only handle a document with
a max height of 750k pixel we transform the range to the virtual size we
have. For example with 40 Million items we have a total height of 1.6
Billion pixel which could not be handled by the browser. Even on 6
Million pixel we got some rubberband issues on chrome.